Split the table.js into smaller files
