+++ web/html/src/manager/audit/subscription-matching/subscription-matching-pins.js
@@ -3,7 +3,10 @@
 
 const React = require("react");
 const createReactClass = require('create-react-class');
-const {Table, Column, SearchField, Highlight} = require("components/table");
+const {Table} = require("components/table/Table");
+const {Column} = require("components/table/Column");
+const {SearchField} = require("components/table/SearchField");
+const {Highlight} = require("components/table/Highlight");
 const PopUp = require("components/popup").PopUp;
 const ModalButton = require("components/dialog/ModalButton").ModalButton;
 const UtilComponent =  require("./subscription-matching-util");
