+++ web/html/src/manager/content-management/list-projects/list-projects.js
@@ -1,7 +1,9 @@
 // @flow
 import React, {useEffect} from 'react';
 import {TopPanel} from 'components/panels/TopPanel';
-import {Column, SearchField, Table} from 'components/table';
+import {Column} from 'components/table/Column';
+import {SearchField} from 'components/table/SearchField';
+import {Table} from 'components/table/Table';
 import Functions from 'utils/functions';
 import {LinkButton} from 'components/buttons';
 import {showSuccessToastr} from 'components/toastr/toastr';
