+++ web/html/src/manager/images/image-view.js
@@ -7,7 +7,9 @@ const {AsyncButton, Button} = require("components/buttons");
 const { TopPanel } = require('components/panels/TopPanel');
 const Network = require("utils/network");
 const Utils = require("utils/functions").Utils;
-const {Table, Column, SearchField} = require("components/table");
+const {Table} = require("components/table/Table");
+const {Column} = require("components/table/Column");
+const {SearchField} = require("components/table/SearchField");
 const Messages = require("components/messages").Messages;
 const MessagesUtils = require("components/messages").Utils;
 const {ModalButton} = require("components/dialog/ModalButton");
