+++ web/html/src/manager/notifications/notification-messages.js
@@ -4,7 +4,10 @@
 const {SectionToolbar} = require("components/section-toolbar/section-toolbar");
 const React = require("react");
 const MessageContainer = require("components/messages").Messages;
-const {Table, Column, SearchField, Highlight} = require("components/table");
+const {Table} = require("components/table/Table");
+const {Column} = require("components/table/Column");
+const {SearchField} = require("components/table/SearchField");
+const {Highlight} = require("components/table/Highlight");
 const Network = require("utils/network");
 const Functions = require("utils/functions");
 const Utils = Functions.Utils;
