+++ web/html/src/manager/systems/ssm/ssm-subscribe-channels.js
@@ -9,7 +9,8 @@ const {ActionSchedule} = require("components/action-schedule");
 const Network = require("utils/network");
 const Functions = require("utils/functions");
 const {Messages} = require("components/messages");
-const {Table, Column} = require("components/table");
+const {Table} = require("components/table/Table");
+const {Column} = require("components/table/Column");
 const { BootstrapPanel } = require('components/panels/BootstrapPanel');
 const MessagesUtils = require("components/messages").Utils;
 const {ChannelLink, ActionLink, ActionChainLink, SystemLink} = require("components/links");
