+++ web/spacewalk-web.changes
@@ -1,3 +1,4 @@
+- Cleanup and document the Table component
 - scheme is not allowed for URI of image store's API endpoint (bsc#1165571)
 
 -------------------------------------------------------------------
